# openapi.yaml
openapi: 3.1.0
info:
  title: OT Toolkit Backend API
  version: 1.0.0
  summary: HTTP projection of the frontend-facing ToolkitService contract.
  description: |
    This specification describes the backend contract expected by the OT Toolkit
    frontend. The desktop application injects HttpToolkitService, which maps
    these HTTP operations back to the same ToolkitService interface used by the
    UI. The FastAPI server uses the seeded MockToolkitService for reference and
    calculation data, with SQLAlchemy persistence for users, tokens, and favorites.

    Reference, discovery, and engineering operations require no authentication.
    Favorites are scoped to a database user and require an opaque bearer token.
    Implementations must remain usable offline; the localhost server below is
    illustrative rather than mandatory.
servers:
  - url: http://127.0.0.1:8000/api/v1
    description: Optional local backend adapter
security: []
...
paths:
...
  /technologies:
    get:
      tags: [Reference]
      operationId: listTechnologies
      summary: List technologies
      description: Implements ToolkitService.list_technologies.
      security: []
      x-authentication: none
      parameters:
        - $ref: '#/components/parameters/CategoryQuery'
      responses:
        '200':
          description: Technologies sorted by display name.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Technology'
        '400':
          $ref: '#/components/responses/BadRequest'
  /technologies/{technologyId}:
    get:
      tags: [Reference]
      operationId: getTechnology
      summary: Get one technology
      description: Implements ToolkitService.get_technology.
      security: []
      x-authentication: none
      parameters:
        - $ref: '#/components/parameters/TechnologyId'
      responses:
        '200':
          description: Complete technology reference record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Technology'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /ports:
    get:
      tags: [Reference]
      operationId: listPorts
      summary: List or search common OT ports
      description: Implements ToolkitService.list_ports.
      security: []
      x-authentication: none
      parameters:
        - $ref: '#/components/parameters/OptionalQuery'
      responses:
        '200':
          description: Matching port records.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PortEntry'
  /glossary:
    get:
      tags: [Reference]
      operationId: listGlossary
      summary: List or search glossary entries
      description: Implements ToolkitService.list_glossary.
      security: []
      x-authentication: none
      parameters:
        - $ref: '#/components/parameters/OptionalQuery'
      responses:
        '200':
          description: Matching glossary entries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GlossaryEntry'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque eight-hour token returned by POST /auth/token.
  parameters:
    TechnologyId:
      name: technologyId
      in: path
      required: true
      description: Stable technology ID from the curated reference data.
      schema:
        type: string
        minLength: 1
    CategoryQuery:
      name: category
      in: query
      required: false
      description: Exact category name. Omit to list all technologies.
      schema:
        type: string
    OptionalQuery:
      name: q
      in: query
      required: false
      description: Case-insensitive free-text filter. Empty or omitted returns all records.
      schema:
        type: string
        default: ''
  responses:
    BadRequest:
      description: Invalid input.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Requested technology or item was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          examples: [invalid_input]
        message:
          type: string
        field:
          type: [string, 'null']
...
    Technology:
      type: object
      additionalProperties: false
      required: [id, name, acronym, category, classification, overview]
      properties:
        id: { type: string }
        name: { type: string }
        acronym: { type: string }
        category: { type: string }
        classification: { type: string }
        overview: { type: string }
        runs_over: { $ref: '#/components/schemas/StringList' }
        related: { $ref: '#/components/schemas/StringList' }
        architecture: { type: string }
        physical_medium: { type: string }
        topology: { type: string }
        speed: { type: string }
        addressing: { type: string }
        ports: { $ref: '#/components/schemas/StringList' }
        files: { $ref: '#/components/schemas/StringList' }
        typical_devices: { $ref: '#/components/schemas/StringList' }
        applications: { $ref: '#/components/schemas/StringList' }
        key_concepts: { $ref: '#/components/schemas/StringList' }
        advantages: { $ref: '#/components/schemas/StringList' }
        limitations: { $ref: '#/components/schemas/StringList' }
        faults: { $ref: '#/components/schemas/StringList' }
        troubleshooting: { $ref: '#/components/schemas/StringList' }
        organization: { type: string }
        standard: { type: string }
        sources: { $ref: '#/components/schemas/StringList' }
    StringList:
      type: array
      items:
        type: string
      default: []
    PortEntry:
      type: object
      additionalProperties: false
      required: [port, transport, name, technology, purpose]
      properties:
        port: { type: string, examples: ['44818'] }
        transport: { type: string, examples: [TCP/UDP] }
        name: { type: string, examples: [EtherNet/IP] }
        technology: { type: string, examples: [CIP] }
        purpose: { type: string }
    GlossaryEntry:
      type: object
      additionalProperties: false
      required: [term, definition, category, related, technology_id]
      properties:
        term: { type: string }
        definition: { type: string }
        category: { type: string }
        related: { $ref: '#/components/schemas/StringList' }
        technology_id:
          type: [string, 'null']
...